win32: update source code to cross compilation
* win32: update all API with the Win32 export symbols methods
  __delspec(dllexport) and __delspec(dllimport) to allow the use of new
  compiler and cross compilation. Add DllDefines.hpp files into all
  dynamic library to define exportation or importation macros.
* win32: append gspawn-win32-helper.exe and
  gspawn-win32-helper-console.exe to permit the use of Glib::spawn
  functions.
* package: fix a bad waitpid use. If the child process is ended correctly
  do not show any message from the errno variable.
* gvle: fix memory leaks in the management of the Gtk::Menu and
  Gtk::MenuItem when using submenu in TreeViewValue and ValuesTreeView.

+++ src/vle/geometry/Point.hpp
@@ -151,7 +151,7 @@ namespace vle { namespace geometry {
          */
         inline Class distance(const Point<n, Class>& p) const
         {
-            return (Class)sqrt((double)distanceSquared(p));
+            return (Class)std::sqrt(distanceSquared(p));
         }
 
         /**
